apps/api/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-sqlalchemy>=3.1
python-dotenv>=1.0
httpx>=0.27
